mock_backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pyyaml>=6.0
httpx>=0.27
